refactor(Alert,Sidebar): shared Action type and StatusTheme
- Move the action type to src/components/shared/action.ts as a
  family-neutral Action (+ ActionContext, mergeActionProps). Alert and
  SidebarCard both use it; the root barrel exports the neutral names.
  AlertAction/AlertActionContext are removed from the public surface.
- AlertTheme and SidebarCardTheme are now aliases of StatusTheme, the
  single source of truth. StatusTheme is exported from the root barrel.
- The 8 status icons leave the public frappe-ui/icons subpath; only
  statusIcon.ts uses them, via direct SFC imports.
