+++ icons/index.ts
@@ -2,16 +2,6 @@ export { default as CircleCheckIcon } from './CircleCheckIcon.vue'
 export { default as DownSolidIcon } from './DownSolidIcon.vue'
 export { default as GreenCheckIcon } from './GreenCheckIcon.vue'
 
-// Status icons (espresso 2.0 Figma glyphs; solid = Alert, line = SidebarCard)
-export { default as AlertCircleSolidIcon } from './AlertCircleSolidIcon.vue'
-export { default as SuccessSolidIcon } from './SuccessSolidIcon.vue'
-export { default as AlertTriangleSolidIcon } from './AlertTriangleSolidIcon.vue'
-export { default as CloseCircleSolidIcon } from './CloseCircleSolidIcon.vue'
-export { default as AlertCircleLineIcon } from './AlertCircleLineIcon.vue'
-export { default as SuccessLineIcon } from './SuccessLineIcon.vue'
-export { default as AlertTriangleLineIcon } from './AlertTriangleLineIcon.vue'
-export { default as CloseCircleLineIcon } from './CloseCircleLineIcon.vue'
-
 // Frappe Icons
 export { default as HelpIcon } from './HelpIcon.vue'
 export { default as LightningIcon } from './LightningIcon.vue'
